Create Adjusted_Momentum.R
This is an adapted clone of the Adjusted Momentum R script from the Systematic Investor blog. Specific alterations to the original R script includes the addition of the SP-500 buy and hold benchmark to the list of models. Please note that this script requires the SIT library to be installed to run. Please see the Systematic Investor blog for instruction on the varies ways to install or import this library.