Create Momentum_and_Markowitz.R
This script is slightly modified from that posted on the systematicinvestor.github.io webpage. Specific modifications where to the 
US.CASH asset composition and the kellerCLAfun was loader from a source file.

The Momentum and Markowitz: A Golden Combination (2015) by Keller, Butler, Kipnis paper is a review of practitioner’s tools to make mean variance optimization portfolio a viable solution. In particular, authors suggest and test:

     1. adding maximum weight limits and
     2. adding target volatility constraint

to control solution of mean variance optimization. 

This script looks at the results for the 8 asset universe:

     1. S&P 500
     2. EAFE
     3. Emerging Markets
     4. US Technology Sector
     5. Japanese Equities
     6. 10-Year Treasuries
     7. T-Bills
     8. High Yield Bonds